Add files via upload
This code aims to calculate simple moving averages (SMA) for different periods in a financial time series data and then calculate the percentage distances between the current price and these moving averages. Additionally, it generates histograms of the distributions of these percentage distances for visualization, as well as plotting the price and moving averages in candlestick format.