Add files via upload
This code was developed to download historical closing price data for the IBOVESPA (^BVSP) and S&P 500 (^GSPC) indices, normalize the prices, and plot the data in a visually attractive manner. The goal is to compare the performance of these indices over time in a clear and intuitive way, eliminating visual distortions of the Data.